TimeSeriesDAODynamoDB superclass. Much refactoring. Wow.
